refactor: extract AI error types and SSRF utilities into dedicated modules
Move AiAnalysisError, mapSdkError, and toSafeAiMessage into aiErrors.ts and extract validateEndpointUrl, createPinnedFetch, and resolvePinnedFetch into aiSsrf.ts to slim aiCaller.ts and improve separation of concerns.
